% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{scad}
\alias{scad}
\title{sj0715
Performs scad}
\usage{
scad(lambda1, lambda2, gamma, diag, X, r, thr, x, yhat, trace, maxiter)
}
\arguments{
\item{lambda1}{lambda}

\item{lambda2}{lambda}

\item{X}{genotype Matrix}

\item{r}{correlations}

\item{thr}{threshold}

\item{x}{beta coef}

\item{yhat}{A vector}

\item{trace}{if >1 displays the current iteration}

\item{maxiter}{maximal number of iterations}
}
\value{
conv
}
\description{
sj0715
Performs scad
}
\keyword{internal}
