% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scadR.R
\name{scadR}
\alias{scadR}
\title{Scad using summary statistics}
\usage{
scadR(
  lambda1,
  lambda2 = 0,
  gamma = 3.7,
  X,
  b,
  thr = 1e-04,
  trace = 0,
  maxiter = 10000,
  blocks = NULL,
  x = NULL
)
}
\description{
Coordinate descent algorithm to solve: 
0.5 x'X'Xx - x'b + 0.5 scad_penalty + 0.5 lambda2 ||x||_2^2 
Function to get scad solutions given X, a reference panel, and
b, regression coefficients, the correlation coefficient r in article
}
\keyword{internal}
